requirements.txt
boto==2.48.0
boto3==1.7.9
botocore==1.10.9
dj-database-url==0.4.2
Django==1.11
django-crispy-forms==1.6.1
django-extensions==2.0.7
django-markdown-deux==1.0.5
django-pagedown==1.0.4
django-storages==1.6.6
djangorestframework==3.8.2
docutils==0.14
gunicorn==19.7.1
jmespath==0.9.3
markdown2==2.3.5
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
python-dateutil==2.7.2
pytz==2017.2
s3transfer==0.1.13
six==1.11.0

